Update registerGridData to handle updates of grid data
Mesh motion and adapativity are handled in TIOGA through repeated calls
to registerGridData. This change allows it to detect the diferent modes
of operation: adding a new mesh block vs. updating the data of an
existing mesh block.

A separate method will be necessary to delete/replace an already
registered mesh block within the data structure.