Fix handling of cellGID field in MeshBlock for large meshes
cellGID is a uint64_t datatype that was being truncated to a 4-byte in
in the MeshBlock::getReceptorInfo method, this was fine for small meshes
but does not work for meshes greater than 2.4B elements. This commit
fixes this by handling the 64-bit data across 2 4-byte integer entries
within the int data structures.