Fix AMR linear interpolation for multiple patches
Current implementation did not use the ghost cell information for the patches,
as such when the donor cell occurred at the patch boundary it used a truncated
stencil for interpolation leading to errors in the interpolation. This commit
modifies the linear interpolation logic to use the fringe information to use an
expanded patch to do the interpolations.

The implementation will still incur an error if the user sets the number of
ghost points to zero, as the interpolation will have to resort to
truncated/one-sided interpolation in that case.