Generalize AMR patch support in TIOGA
Update CartGrid interface to support two objectives:

Ensure that there is a valid AMRMeshInfo object on host and device even when
using the legacy AMR mesh registration interface.

Introduce a new `tioga::preprocess_amr_data` that will ensure that a valid
CartGrid instance exists on all MPI ranks visible to TIOGA even if the AMR
solver is running on a subset of MPI processes.